Delete confirmation dialog

+++ EmployeeManagement.Web/Pages/DisplayEmployeeBase.cs
@@ -1,4 +1,5 @@
-﻿using EmployeeManagement.Models;
+﻿using EmployeeManagement.Components;
+using EmployeeManagement.Models;
 using EmployeeManagement.Web.Services;
 using Microsoft.AspNetCore.Components;
 using System;
@@ -8,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace EmployeeManagement.Web.Pages
 {
-    public class DisplayEmployeeBase:ComponentBase
+    public class DisplayEmployeeBase : ComponentBase
     {
         [Parameter]
         public Employee Employee { get; set; }
@@ -24,12 +25,30 @@ namespace EmployeeManagement.Web.Pages
         public IEmployeeService EmployeeService { get; set; }
         [Inject]
         public NavigationManager NavigationManager { get; set; }
-        protected async Task Delete_Click()
+
+        protected ConfirmBase DeleteConfirmation {get; set;}
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
-            //NavigationManager.NavigateTo("/",true);
+            DeleteConfirmation.Show();
         }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+            }
+        }
+
+        //protected async Task Delete_Click()
+        //{
+        //    await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+        //    await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+        //    //NavigationManager.NavigateTo("/",true);
+        //}
+
+
         protected async Task CheckBoxChanged(ChangeEventArgs e)
         {
             IsSelected =(bool) e.Value;
