+++ EmployeeManagement.Web/Pages/EditEmployeeBase.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using EmployeeManagement.Components;
 using EmployeeManagement.Models;
 using EmployeeManagement.Web.Models;
 using EmployeeManagement.Web.Services;
@@ -30,6 +31,7 @@ namespace EmployeeManagement.Web.Pages
         [Parameter]
         public string Id { get; set; }
         public string PageTitle { get; set; }
+        protected ConfirmBase DeleteConfirmation { get; set; }
         protected async override Task OnInitializedAsync()
         {
             int.TryParse(Id, out int employeeId);
@@ -85,10 +87,25 @@ namespace EmployeeManagement.Web.Pages
             }
         }
 
-        protected async Task Delete_Click()
+        
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            NavigationManager.NavigateTo("/");
+            DeleteConfirmation.Show();
         }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                NavigationManager.NavigateTo("/");
+                //await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+            }
+        }
+        //protected async Task Delete_Click()
+        //{
+        //    await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+        //    NavigationManager.NavigateTo("/");
+        //}
     }
 }
